refresh-channel-stats: explicit ::bigint cast — fixes overflow on channels with >2.1B views
Surfaced on the 10-channel listicle render attempt (job 113 trigger):
every "whale" channel failed loadChannel with:
  [loadChannel] refresh failed for UCX...: value "128672607412" is out
  of range for type integer

Affected:
  MrBeast        128B views
  Like Nastya    121B
  BANGTANTV       28B
  Toys & Colors  119B
  Mark Rober      17B
  Movieclips      70B

Root cause: the UPDATE used `$N > 0` which made Postgres bind $N as
INTEGER (not BIGINT) at the comparison. total_views is BIGINT in the
table schema, but the parameter type was determined by the comparison,
not the destination column. Values > 2^31 (2.1B) overflowed at bind
time before any UPDATE happened.

Fix: explicit `::bigint` cast on $2 (subscriber_count — MrBeast et al
are below INT4 max today but future-proof) and $4 (total_views) at
both the comparison AND the assignment. video_count stays INTEGER
(no channel has >2.1B videos).

loadChannel's try/catch was already swallowing the error and falling
back to the existing DB row, but the existing row was stale, so the
listicle's narration referenced outdated subscriber counts.

+++ lib/content-gen/refresh-channel-stats.ts
@@ -134,11 +134,17 @@ export async function refreshChannelStats(
 
   // 4. Persist. Only overwrite columns when API returns >0 (defensive against
   //    YT occasionally returning zero counts during rolling updates).
+  //    EXPLICIT ::bigint casts on subscriber_count and total_views — these
+  //    can exceed INT4 max (2.1B) for whales like MrBeast (128B views). Without
+  //    the cast, Postgres types the `$N > 0` comparison as INTEGER and
+  //    overflows at bind time with "value X is out of range for type integer"
+  //    (observed 2026-06-10 on the 10-channel listicle render — every channel
+  //    with >2.1B views failed loadChannel).
   await pool.query(
     `UPDATE niche_spy_channels SET
-       subscriber_count   = CASE WHEN $2 > 0 THEN $2 ELSE subscriber_count END,
+       subscriber_count   = CASE WHEN $2::bigint > 0 THEN $2::bigint ELSE subscriber_count END,
        video_count        = CASE WHEN $3 > 0 THEN $3 ELSE video_count END,
-       total_views        = CASE WHEN $4 > 0 THEN $4 ELSE total_views END,
+       total_views        = CASE WHEN $4::bigint > 0 THEN $4::bigint ELSE total_views END,
        stats_refreshed_at = $5
      WHERE channel_id = $1`,
     [channelId, subscriberCount, videoCount, viewCount, refreshedAt],
